Adjust casts and format code

+++ src/hal/sensor/hyt939.cpp
@@ -24,7 +24,7 @@ std::optional<Hyt939Data> hyt939Measure()
       {
          return {};
       }
-   } while(buffer[0] & ~0xbf);
+   } while (buffer[0] & ~0xbf);
 
    /* get data */
    if (i2cMasterRead(HYT939_ADDR, buffer, sizeof(buffer)) != I2cStatus::SUCCESS)
@@ -33,10 +33,12 @@ std::optional<Hyt939Data> hyt939Measure()
    }
 
    Hyt939Data data;
-   const uint16_t humi_raw = ((buffer[0]<<8) | buffer[1]) & 0x3fff;
-   const uint16_t temp_raw = ((buffer[2]<<6) | (buffer[3]>>2)) & 0x3fff;
-   data.humidity = static_cast<uint32_t>(humi_raw) * 100UL / ((2<<13) - 1);
-   const uint16_t temp_tmp = static_cast<uint32_t>(temp_raw) * 1650UL / ((2<<13) - 1);
+   const uint16_t humi_raw = ((buffer[0] << 8) | buffer[1]) & 0x3fff;
+   const uint16_t temp_raw = ((buffer[2] << 6) | (buffer[3] >> 2)) & 0x3fff;
+   data.humidity =
+      (static_cast<uint32_t>(humi_raw) * 100UL) / (static_cast<uint32_t>(2 << 13) - 1UL);
+   const uint16_t temp_tmp =
+      static_cast<uint32_t>(temp_raw) * 1650UL / (static_cast<uint32_t>(2 << 13) - 1UL);
    data.temperature = static_cast<int16_t>(temp_tmp) - 400;
 
    return data;
